Initial commit
This is the rudimentary Python script to manage video files from the
helmet cam.
It will rename video files based on the parent directory's last
filename, move the file to the parent directory, and log its actions in
a log file.
To do: Properly list the history of the files. (Can you do this in Git?)